Update de_DE.lang
Fixed some Typos
Hähnchennugget (cockerelnugget) -> Hühnchennugget (chickennugget)
Container -> Behälter
Removed a bunch of unneccessary or wrong commas
Order -> Ordnung (I interpret it as the opposite of Entropy, and not as order to do something)
Note: I didn't check tc.research_* yet, as it's getting late. I'll check them the next days, if/when I got the time for it.